pyproject.toml
[project]
name = "spikeinterface"
version = "0.101.1"
authors = [
  { name="Alessio Buccino", email="alessiop.buccino@gmail.com" },
  { name="Samuel Garcia", email="sam.garcia.die@gmail.com" },
]
description = "Python toolkit for analysis, visualization, and comparison of spike sorting output"
readme = "README.md"
requires-python = ">=3.8,<4.0"
classifiers = [
    "Programming Language :: Python :: 3 :: Only",
    "License :: OSI Approved :: MIT License",
    "Intended Audience :: Science/Research",
    "Operating System :: POSIX :: Linux",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: MacOS",
    "Operating System :: OS Independent"
]


dependencies = [
    "numpy>=1.20, <2.0",  # 1.20 np.ptp, 1.26 might be necessary for avoiding pickling errors when numpy >2.0
    "threadpoolctl>=3.0.0",
    "tqdm",
    "zarr>=2.16,<2.18",
    "neo>=0.13.0",
    "probeinterface>=0.2.23",
    "packaging",
]

[build-system]
requires = ["setuptools>=62.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true
package-data = {"spikeinterface.sorters" = ["**/*.m", "**/*.prm", "**/*.params", "**/*.yaml"]}

[tool.setuptools.exclude-package-data]
spikeinterface = ["**/tests/test_*"]

[tool.setuptools.packages.find]
where = ["src"]
include = ["spikeinterface*"]
namespaces = false
exclude = ["spikeinterface.*.tests"]

[tool.black]
line-length = 120

[project.urls]
homepage = "https://github.com/SpikeInterface/spikeinterface"
repository = "https://github.com/SpikeInterface/spikeinterface"
documentation = "https://spikeinterface.readthedocs.io/"
changelog = "https://spikeinterface.readthedocs.io/en/latest/whatisnew.html"


[project.optional-dependencies]

extractors = [
    "MEArec>=1.8",
    "pynwb>=2.6.0",
    "hdmf-zarr>=0.5.0",
    "pyedflib>=0.1.30",
    "sonpy;python_version<'3.10'",
    "lxml", # lxml for neuroscope
    "scipy",
    "ONE-api>=2.7.0", # alf sorter and streaming IBL
    "ibllib>=2.36.0", # streaming IBL
    "pymatreader>=0.0.32", # For cell explorer matlab files
    "zugbruecke>=0.2; sys_platform!='win32'", # For plexon2
]

streaming_extractors = [
    "ONE-api>=2.7.0", # alf sorter and streaming IBL
    "ibllib>=2.36.0", # streaming IBL
    # Following dependencies are for streaming with nwb files
    "pynwb>=2.6.0",
    "fsspec",
    "aiohttp",
    "requests",
    "hdmf-zarr>=0.5.0",
    "remfile",
    "s3fs"
]


preprocessing = [
    "scipy",
]


full = [
    "h5py",
    "pandas",
    "xarray",
    "scipy",
    "scikit-learn",
    "networkx",
    "distinctipy",
    "matplotlib>=3.6", # matplotlib.colormaps
    "cuda-python; platform_system != 'Darwin'",
    "numba",
]

widgets = [
    "matplotlib",
    "ipympl",
    "ipywidgets",
    "sortingview>=0.12.0",
]

icms-plasticity = [
    "black",
    "pre-commit",
    "seaborn",
    "tkfilebrowser",
    "python-pptx",
    "mountainsort5",
    "umap-learn"
]

qualitymetrics = [
    "scikit-learn",
    "scipy",
    "pandas",
    "numba",
]

test_core = [
    "pytest",
    "pytest-dependency",
    "psutil",

    # for github test : probeinterface and neo from master
    # for release we need pypi, so this need to be commented
    "probeinterface @ git+https://github.com/SpikeInterface/probeinterface.git",
    "neo @ git+https://github.com/NeuralEnsemble/python-neo.git",
]

test_extractors = [
    # Functions to download data in neo test suite
    "pooch>=1.8.2",
    "datalad>=1.0.2",
    "probeinterface @ git+https://github.com/SpikeInterface/probeinterface.git",
    "neo @ git+https://github.com/NeuralEnsemble/python-neo.git",
]

test_preprocessing = [
    "ibllib>=2.36.0", # for IBL
    "torch",
]


test = [
    "pytest",
    "pytest-dependency",
    "pytest-cov",

    "xarray",
    "huggingface_hub",

    # preprocessing
    "ibllib>=2.36.0", # for IBL

    # streaming templates
    "s3fs",

    # tridesclous
    "numba",
    "hdbscan>=0.8.33",  # Previous version had a broken wheel

    # for sortingview backend
    "sortingview",


    ## install tridesclous for testing ##
    "tridesclous>=1.6.8",

    ## sliding_nn
    "pymde",
    "torch",
    "pynndescent",

    # for github test : probeinterface and neo from master
    # for release we need pypi, so this need to be commented
    "probeinterface @ git+https://github.com/SpikeInterface/probeinterface.git",
    "neo @ git+https://github.com/NeuralEnsemble/python-neo.git",
]

docs = [
    "Sphinx",
    "sphinx_rtd_theme",
    "sphinx-gallery",
    "sphinx-design",
    "numpydoc",
    "ipython",

    # for notebooks in the gallery
    "MEArec", # Use as an example
    "pandas", # in the modules gallery comparison tutorial
    "hdbscan>=0.8.33",   # For sorters spykingcircus2 + tridesclous
    "numba", # For many postprocessing functions
    "xarray", # For use of SortingAnalyzer zarr format
    "networkx",
    # Download data
    "pooch>=1.8.2",
    "datalad>=1.0.2",

    # for release we need pypi, so this needs to be commented
    "probeinterface @ git+https://github.com/SpikeInterface/probeinterface.git",  # We always build from the latest version
    "neo @ git+https://github.com/NeuralEnsemble/python-neo.git",  # We always build from the latest version

]

dev = [
    "spikeinterface[test]",
    "spikeinterface[test_core]",
    "spikeinterface[docs]",
    "black",
    "pre-commit",
]

[tool.pytest.ini_options]
markers = [
    "core",
    "generation",
    "extractors",
    "preprocessing",
    "postprocessing",
    "qualitymetrics",
    "sorters",
    "sorters_external",
    "sorters_internal",
    "comparison",
    "curation",
    "exporters",
    "widgets",
    "sortingcomponents",
    "streaming_extractors: extractors that require streaming such as ross and fsspec",
]
filterwarnings =[
    'ignore:.*distutils Version classes are deprecated.*:DeprecationWarning',
    'ignore:.*the imp module is deprecated in favour of importlib.*:DeprecationWarning',
]
